requirements.txt
tensorflow>=2.8.0
streamlit>=1.18.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.2
scikit-learn>=1.0.2
Pillow>=9.0.0
pandas>=1.4.0
requests>=2.27.0